database

+++ src/core/models/header.cr
@@ -1,6 +1,6 @@
 module ::Sushi::Core::Models
   alias Header = NamedTuple(
-          index: UInt32,
+          index: Int64,
           nonce: UInt64,
           prev_hash: String,
           merkle_tree_root: String,
